requirements.txt
python-dotenv
gradio
faiss-cpu
sentence-transformers
transformers
accelerate
pymupdf
ollama
google-generativeai
beautifulsoup4
python-docx
fastapi
uvicorn
python-multipart
passlib>=1.7.4
python-jose[cryptography] 
sqlalchemy
mysqlclient
PyJWT
pymysql
pydantic[email]
bcrypt==3.2.2
